+++ WarehouseManagement/Models/Product.cs
@@ -8,10 +8,10 @@ namespace WarehouseManagement.Models
 {
     public class Product
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public string Name { get; set; }
         public int Price { get; set; }
         public string Description { get; set; }
-        public int Weight { get; set; }
+        public double Weight { get; set; }
     }
 }
